fix: evaluate memory page limit without overflow

+++ crates/telomere/src/common.rs
@@ -1056,7 +1056,7 @@ impl ConstExpr {
     }
 }
 pub const PAGE_SIZE: usize = 64 * 1024;
-pub const PAGE_SIZE_MAX: usize = 4 * 1024 * 1024 * 1024 / PAGE_SIZE;
+pub const PAGE_SIZE_MAX: usize = (4_u64 * 1024 * 1024 * 1024 / PAGE_SIZE as u64) as usize;
 
 pub struct ExecuteContext<'a> {
     pub stack: &'a mut Stack,
